fix(tools): resolve design-system audit findings in tools catalog
